handle `""` earlier, without setting up the loop
to count \s.  That turns out to be an improvement
for my input, even though it doesn't contain a lot
number of "" per record (typically around 8).

$ time gjq  '.' <100k_json_records.json >/dev/null
real    0m17.581s
user    0m16.896s
sys     0m0.628s

before this it took real   0m17.778s. It's not much but
the improvement is consistent across runs.